+++ src/note/api-handlers.go
@@ -20,6 +20,7 @@ func Handler(res http.ResponseWriter, req *http.Request) {
 }
 
 // TODO: this function is flawed. use code from userhandler and implement it here
+
 // APIHandleByID : handles the requests to notes with a certain id
 func APIHandleByID(res http.ResponseWriter, req *http.Request) {
 	if req.Method == http.MethodGet {
